Reduced pyramid precision from float32 to float16 for faster reads/writes to disk

+++ src/algorithm/image_storage.py
@@ -15,14 +15,14 @@ class ImageStorageHandler:
     def __init__(self):
         return
 
-    # Write laplacian pyramid of image to archive on disk
+    # Write laplacian pyramid of image to archive on disk (as float16)
     def write_laplacian_pyramid_to_disk(self, laplacian_pyramid, root_dir):
         file_handle, tmp_file = tempfile.mkstemp(".npz", None, root_dir.name)
         dictionary = {}
 
         # Append pyramid levels to dictionary
         for i, pyramid_level in enumerate(laplacian_pyramid):
-            dictionary["Laplacian_" + str(i)] = pyramid_level
+            dictionary["Laplacian_" + str(i)] = pyramid_level.astype(np.float16)
 
         # Write to uncompressed archive
         np.savez(tmp_file, **dictionary)
@@ -31,7 +31,7 @@ class ImageStorageHandler:
         os.close(file_handle)
         return tmp_file
 
-    # Load laplacian pyramid from archive on disk
+    # Load laplacian pyramid from archive on disk (as float32)
     def load_laplacian_pyramid(self, image_archive_filename):
         archive = np.load(image_archive_filename, allow_pickle=False)
 
@@ -39,6 +39,6 @@ class ImageStorageHandler:
         laplacian_pyr_keys = sorted(archive.files, key=utilities.int_string_sorting)
         laplacian_pyr = List()
         for key in laplacian_pyr_keys:
-            laplacian_pyr.append(archive[key])
+            laplacian_pyr.append(archive[key].astype(np.float32))
 
         return laplacian_pyr
